% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_hexbin.R
\name{base_hexbin}
\alias{base_hexbin}
\title{Title}
\usage{
base_hexbin(df, seed)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
NOT DONE: Making sure functions from other packages are labeled: 'package::function()'
}
