% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram.R
\name{cartogram}
\alias{cartogram}
\title{Creates an interactive cartogram map using the viridis palette}
\usage{
cartogram(data, weight, variable, geometry, base = NULL, ...)
}
\arguments{
\item{data}{A data set}

\item{weight}{The variable used to scale the cartogram}

\item{variable}{The variable corresponding to the palette}

\item{geometry}{The geometry/multipolygon variable}

\item{base}{Adds a base map to the background; the default is NULL}

\item{...}{}
}
\value{
An interactive cartogram map
}
\description{
Creates an interactive cartogram map using the viridis palette
}
\examples{
cartogram(states, "Population", "Vegetables", geometry, NULL)
cartogram(states, "Population", "Vegetables", geometry, "var")
}
