% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Creates an interactive choropleth map}
\usage{
choropleth(data, variable, geometry, palette = "viridis", ...)
}
\arguments{
\item{data}{A data set with a geometry/multipolygon variables}

\item{variable}{A variable whose amount will color the map}

\item{geometry}{A geometry/multipolygon variable of the region of interest (i.e. city, state, country, etc.)}

\item{palette}{A palette that corresponds to the variable; default is "viridis."}

\item{...}{}
}
\value{
An interactive choropleth map
}
\description{
Creates an interactive choropleth map
}
\examples{
choropleth(states, "Population", geometry, "viridis")
}
