% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_polygons.R
\name{label_polygons}
\alias{label_polygons}
\title{Title}
\usage{
label_polygons(shapefile_path, shapefile, labels, shapefile_key, labels_key)
}
\arguments{
\item{labels_key}{}
}
\description{
Title
}
\examples{
NOT DONE: Making sure functions from other packages are labeled: 'package::function()'
}
