% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointmap.R
\name{pointmap}
\alias{pointmap}
\title{Creates an interactive map with points using longitude and latitude}
\usage{
pointmap(
  data_set,
  longitude_var,
  latitude_var,
  set_longitude,
  set_latitude,
  popups = NULL,
  icon_filepath = NULL,
  icon_width = 15,
  icon_height = 15,
  zoom_min = 12,
  set_zoom = 13,
  zoom_max = 19,
  map_tile = leaflet::providers$Stamen.Terrain,
  point_color = "red",
  point_radius = 10,
  point_inopacity = 0.2,
  point_outopacity = 0.4,
  user_pal = NULL,
  user_var = NULL
)
}
\arguments{
\item{data_set}{A data set with longitude and latitude variables}

\item{longitude_var}{Variable containing longitude geographic coordinate}

\item{latitude_var}{Variable containing latitude geographic coordinate}

\item{set_longitude}{Longitude of general region (i.e. city, state, country)}

\item{set_latitude}{Latitude of general region (i.e. city, state, country)}

\item{popups}{Text that will appear when a point is pressed (i.e. date, etc.)}

\item{icon_filepath}{Filepath of icon for point}

\item{icon_width}{Width of icon}

\item{icon_height}{Height of icon}

\item{zoom_min}{Minimum zoom}

\item{set_zoom}{General zoom level}

\item{zoom_max}{Maximum zoom}

\item{map_tile}{Map base using leaflet providers, options:
leaflet::providers$Stamen.Toner,
leaflet::providers$Stamen.Watercolor}

\item{point_color}{Color of points on map}

\item{point_radius}{Radius of points on map}

\item{point_inopacity}{Opacity of point fill, default = 0.2}

\item{point_outopacity}{Opacity of point outline, default = 0.4}

\item{user_pal}{Palette for points, must match number of factors in variables}

\item{user_var}{Variable to change hue of points to, corresponds to palette}
}
\value{
An interactive map with points
}
\description{
Creates an interactive map with points using longitude and latitude
}
\examples{
pointmap(data = crash_data,
longitude_var = crash_data$lon,
latitude_var = crash_data$lat
set_longitude = -71.110558,
set_latitude = 42.3736)

pointmap(data = crash_data,
longitude_var = crash_data$lon,
latitude_var = crash_data$lat,
set_longitude = -71.110558,
set_latitude = 42.3736,
popups = crash_data$crash_date,
user_pal = c("#003f5c", "#2f4b7c","#665191",
             "#a05195","#d45087","#f95d6a",
             "#ff7c43","#ffa600"),
user_var = crash_data$year)
}
