% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_hexbin.R
\name{seed_hexbin}
\alias{seed_hexbin}
\title{Title}
\usage{
seed_hexbin(df)
}
\arguments{
\item{df}{}
}
\description{
Title
}
\examples{
NOT DONE: Making sure functions from other packages are labeled: 'package::function()'
}
