% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{tessellate}
\alias{tessellate}
\title{Fits hexagon tessellations to geographical multipolygon data as base for hexbin maps}
\usage{
tessellate(data, fill, geometry, n, ...)
}
\arguments{
\item{data}{Dataset with geometry/multipolygon data}

\item{fill}{Variable used to fill the hexagons}

\item{geometry}{Variable that gives geometry/multipolygon data}

\item{n}{Number of seed maps to generate}

\item{...}{}
}
\value{
Possible hexagon tessellations as base for hexbin maps
}
\description{
Fits hexagon tessellations to geographical multipolygon data as base for hexbin maps
}
\examples{
seed_hexbin(data = states, fill = "Vegetables", geometry = geometry, n = 6)
}
