% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{hexbin}
\alias{hexbin}
\title{Creates an interactive hexbin map using the viridis palette}
\usage{
hexbin(
  data,
  fill,
  labels,
  geometry,
  palette = "viridis",
  seed = NULL,
  base = NULL,
  ...
)
}
\arguments{
\item{data}{Dataset with geometry/multipolygon data}

\item{fill}{Variable used to fill the hexagons}

\item{labels}{Variable used to label the hexagons}

\item{geometry}{Variable that gives geometry/multipolygon data}

\item{palette}{Palette for filling hexagons; default is Viridis}

\item{seed}{Number corresponding to seed map chosen using tessellate()}

\item{...}{}
}
\value{
An interactive hexbin map
}
\description{
Creates an interactive hexbin map using the viridis palette
}
\examples{
hexbin(data = states,
fill = "Vegetables",
labels = "STUSPS",
geometry = geometry,
palette = "viridis")
}
