% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessellate.R
\name{tessellate}
\alias{tessellate}
\title{Fits hexagon tessellations to geographical multipolygon data as base for hexbin maps}
\usage{
tessellate(data, geometry, n = 6, ...)
}
\arguments{
\item{data}{Dataset with geometry/multipolygon data}

\item{geometry}{Variable that gives geometry/multipolygon data}

\item{n}{Number of seed maps to generate; default 6}

\item{...}{}
}
\value{
Possible hexagon tessellations as base for hexbin maps
}
\description{
Fits hexagon tessellations to geographical multipolygon data as base for hexbin maps
}
\examples{
tessellate(data = states, geometry = geometry, n = 6)
}
